requirements.txt
python-telegram-bot[job-queue]==21.3
anthropic>=0.34.0
openai>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
